Add db_query CLI with sub commands for querying and filtering database
CLI has been created to allow separate subcommands for specific types
of queries. The general structure of the CLI is: db-query (options
pertaining to all subcommands) DB_NAME [subcommand] (options pertaining
to specific subcommand). Filtering functionality on results has also
been added. These can be invoked with the optional flags following the
query subcommand, such as `--date-before`. Now that the general structure
of the CLI is in place, one can easily add subcommands in future.

New function get_analyses returns a pandas data frame containing all analyses
conducted to date, along with their metadata.

New function filter_dataframe accepts a generic dataframe and filters out
rows that do not match column_name: value pairs as specified in a dict.
Still to be determined: the exact filtering approach to be implemented;
should it be performed at the level of the dataframe or should filtering
all be done at a lower level such as when the queries are performed by
sqlalchemy.